Resolves JENKINS-22125
Previously, when the Parameterized Trigger Plugin only passed
StringParameterValues, invalid values could be passed to downstream
jobs, eg passing an invalid value for a Choice Parameter. Now that the
values get converted to their appropriate type, use cases depending on
allowing invalid values to be passed break with an
IllegalArgumentException.

This patch broadens compatibility with existing use cases that are
dependent on passing an invalid value without sacrificing the intended
functionality of the feature.