Add FAILURE_OR_BETTER Trigger Threshold.
This Threshold is missing in compare to original Jenkins Trigger Thresholds.

ALWAYS is not the same, as ALWAYS also triggers the next build if the user abort the running job.

So if you have a Pipeline with 15 Steps and last 10 Steps are connected by ALWAYS, you have to abort 10 jobs in a row.

FAILURE_OR_BETTER is the better choice for this.