Provide an option to skip triggering when a build job is invoked manually.
Add an option which allows the user to choose between executing triggers like they normally do, or to skip them when a user was the cause of the build of this job. This allows users to decide whether or not they want to invoke a whole pipeline of jobs or if they just want to invoke a single job.