allow indicating project name in property files
When triggering multiple builds, it is now possible to wait for
completion on multiple projects and indicate in property files what
project they should trigger a build on.

For example, if we have 5 properties files:

//foo.properties ------------------------
JOB=foo
PARAM=12
// --------------------------------------
//bar-1.properties ----------------------
JOB=bar
PARAM=1
// --------------------------------------
//bar-2.properties ----------------------
JOB=bar
PARAM=2
// --------------------------------------
//baz-1.properties ----------------------
JOB=baz
PARAM=1
// --------------------------------------
//baz-2.properties ----------------------
JOB=baz
PARAM=2
// --------------------------------------

And the following configuration on the parent job:

    Projects to build: "foo, bar, baz"
    Block until the triggered projects finish their builds: true
    For every property file, invoke one build:
        File pattern: *.properties
        Project Name Property: JOB

The result will be the following triggered builds:

    "foo" with params from "foo.properties"
    "bar" with params from "bar-1.properties"
    "bar" with params from "bar-2.properties"
    "baz" with params from "baz-1.properties"
    "baz" with params from "baz-2.properties"

If the "Project Name Property" field is left blank, the behaviour is the
same as before: "For every property file, invoke one build ON ALL
"Projects to build".

Important: the value of "Project Name Property" is NOT passed as a
parameter to the downstream builds.

Signed-off-by: Robin Jarry <robin.jarry@6wind.com>