TriggerBuilder now implements DependencyDeclarer
A couple of other plugins rely on the fact that the
DependencyGraph is setup correctly, e.g. Junit and
Downstream-buildview. This change enables TriggerBuilder,
which is invoked when downstream builds are added as a
build step, to work in the same way as BuildTrigger,
which is used for post build steps.

Change-Id: I38750f5f11529c4fbe7effa969f5ba35b57da3b1