TriggerBuilder now implements DependencyDeclarer
Certain plugins doesn't work when triggering builds
of other projects as build steps, i.e. using TriggerBuilder.
When triggering as a post build step, i.e. using BuildTrigger,
this works. Two examples are the Junit plugin, when aggregating
test results from downstream builds and the downstream build view.

The reason is that BuildTrigger builds up a correct
DependencyGraph and TriggerBuilder does not.

This change remedies this by using the same mechanism as BuildTrigger.