TriggerBuilder now implements DependencyDeclarer (#116)
* TriggerBuilder now implements DependencyDeclarer

A couple of other plugins rely on the fact that the
DependencyGraph is setup correctly, e.g. Junit and
Downstream-buildview. This change enables TriggerBuilder,
which is invoked when downstream builds are added as a
build step, to work in the same way as BuildTrigger,
which is used for post build steps.

Change-Id: I38750f5f11529c4fbe7effa969f5ba35b57da3b1

* Fix issue with builds being triggered twice

All builds triggered through parameterized trigger buildstep
will be triggered twice. Once through the parameterized build
step and once as a post build step through the dependency
graph (see AbstractBuild.cleanUp()).

Most of the times the triggering will happen quick enough,
so both builds lands in the queue and Jenkins will then
merge them.

However, in some circumstances, the first build will start
building before the second one is scheduled. This is when we
see two seperate builds.

Fix issue by allowing the dependency graph to indicate that
Jenkins should not trigger the builds.

Change-Id: Ie66af4ef85ee4f3c9aa9bbe1bba5b231fd325b61

* Added test for verifying a build is only triggered once

* CanDeclare() only needs to check for PromotionProcess

Since we now depends on Jenkins 1.642.3, we know that
the DependencyGraph will exist.