Feature: System property ptp.disableActionViews disables summary
Set Java property "ptp.disableActionViews" (typically Launch Jenkins
with -Dptp.disableActionViews) to stop the listing of "Sub Projects"
on the build and project main pages. This can be of use when other
plugins are used to render the upstream/downstream flow and there is
reduntant information displayed about the triggered builds.