Handle NOT_BUILT and ABORTED as other results
Use case
--------

My use case is a post build script triggering a job which does not need
to fully complete. Since the build step can be cancelled while waiting
or building, it causes the triggering build to be marked as a failure
despite asking to never block.

To fix that, I need the plugin to pass the result of the build step
through BlockingBehaviour. That lets one define how to behave when the
triggered build is NOT_BUILT (it got cancelled from the queue). In my
case I need it to never block.

See: https://phabricator.wikimedia.org/T352319

Solution
--------

There are multiple reasons for a job to not fully complete:

- it is interrupted, an InterruptedException is thrown, this is still
  rethrown and Jenkins will mark the build as ABORTED.

- it can be cancelled from the build queue raising a
  CancellationException. This previously raised an AbortException which
  Jenkins handles by marking the build as a failure. I have changed it to
  a NOT_BUILT result which can be process as other results (addressing
  my use case to have it to never block).

The Jenkins Result class ranks the results as:
- SUCCESS
- UNSTABLE
- FAILURE
- NOT_BUILT
- ABORTED.

The NOT_BUILT and ABORTED results are thus worse than a FAILURE and
would be matched as such in BlockingBehavior mapBuildStepResult() and
mapBuildResult() which both use isWorseOrEqualTo() for comparison.

Add a test testCancelledFromBuildQueue() to cover the
CancellationException() is caught and it results in a SUCCESS (since the
test blocking behavior is to never block).

The ResultConditionTest test covers that BlockingBehavior is able to map
NOT_BUILD and ABORTED since it has two tests explicitly cancelling and
interrupting jobs.

Examples
--------

When a build is ongoing and when aborting it:

  Waiting for the completion of downstream-project
  downstream-project #7 started.
  downstream-project #7 completed. Result was ABORTED
  Build step 'Trigger/call builds on other projects' marked build as failure
  Finished: FAILURE

When it is waiting in the build queue and get cancelled:

  Waiting for the completion of downstream-project
  Not built: downstream-project has been cancelled while waiting in the queue.
  Build step 'Trigger/call builds on other projects' marked build as failure
  Finished: FAILURE